+++ src/main.ts
@@ -4,8 +4,8 @@ import { Container } from 'inversify';
 import { Application } from './app/rest.js';
 import { Service } from './types/service.js';
 import { createRestApplicationContainer } from './app/rest.container.js';
-import { createUserContainer } from './modules/user/user.container.js';
-import { createOfferContainer } from './modules/offer/offer.container.js';
+import { createUserContainer } from './services/user/user.container.js';
+import { createOfferContainer } from './services/offer/offer.container.js';
 
 (async function bootstrap () {
   const mainContainer = Container.merge(
