+++ src/services/offer/offer-service.interface.ts
@@ -1,6 +1,6 @@
 import { DocumentType } from '@typegoose/typegoose';
 import { OfferEntity } from './offer.entity.js';
-import CreateOfferDto from './dto/create-offer.dto.js';
+import CreateOfferDto from './dto/create-offer.dto';
 
 export interface OfferServiceInterface {
   create(dto: CreateOfferDto): Promise<DocumentType<OfferEntity>>;
