+++ src/services/offer/offer.entity.ts
