+++ src/services/user/user-service.interface.ts
@@ -1,5 +1,5 @@
 import { DocumentType } from '@typegoose/typegoose';
-import CreateUserDto from './dto/create-user.dto.js';
+import CreateUserDto from './dto/create-user.dto';
 import { UserEntity } from './user.entity.js';
 
 export interface UserServiceInterface {
