+++ src/services/user/user.container.ts
@@ -1,7 +1,7 @@
 import { Container } from 'inversify';
-import {UserServiceInterface} from "./user-service.interface.js";
-import {Service} from "../../types/service.js";
-import UserService from "./user.service.js";
+import {UserServiceInterface} from './user-service.interface';
+import {Service} from '../../types/service.js';
+import UserService from './user.service.js';
 import { types } from '@typegoose/typegoose';
 import { UserEntity, UserModel } from './user.entity.js';
